Tests: split generic download-CLI tests out of test_download_fsspec.py
Pure move (review: PR #173). tests/test_download_cli.py now holds every test that
drives the meds-extract-download console script and asserts CLI-level contracts —
exit codes, bucket key= semantics, selective interpolation resolution, pkg:// spec
resolution, the reserved sources.dataset_version key, overwrite/fail-fast policy —
where local fsspec mirrors are only the no-network transport vehicle:

  - test_meds_extract_download_cli_end_to_end
  - test_cli_only_resolves_sources_subtree
  - _run_cli (+ _issue_151_spec / _issue_151_env helpers)
  - test_cli_failure_exits_nonzero
  - test_cli_unknown_key_exits_nonzero
  - test_cli_malformed_source_entry_logs_and_exits_nonzero
  - test_cli_no_sources_block_warns_and_exits_zero
  - test_cli_key_selects_bucket_and_appends_common
  - test_cli_unselected_bucket_interpolations_not_resolved
  - test_cli_selected_bucket_interpolation_failure_is_clear
  - test_cli_cross_source_collision_exits_before_any_fetch
  - test_cli_fail_fast_skips_remaining_sources
  - test_cli_pkg_spec_resolution (new in this PR)
  - test_cli_sources_dataset_version_reserved_key (new in this PR)

test_download_fsspec.py keeps the Source-/FsspecSource-level tests
(symlink-escape boundary, include-before-hashing, pooled dispatch window,
memory:// protocol) and is otherwise byte-identical to dev again — this also
evaporates the two docformatter reflow hunks the review flagged: the pinned
docformatter (1.7.5) reflows those two pre-existing docstrings only when content
is appended later in the same file (verified by bisection; a clean dev tree is
stable under pre-commit), so removing the appended tests removes the churn. The
same two docstrings ARE canonicalized by docformatter inside the new file's
layout — visible there as new-file content, not as noise on existing files.
